+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 ### Corrected Bugs
 
+* Fixed bug regarding number of lines in CLIgen history.
+  * The line number given to `cligen_hist_init()` saved one less than the number given.
 * [Using a regular expression with either ^ OR $ breaks the regular expression match logic](https://github.com/clicon/cligen/issues/53)
 * Recursive sets did not work properly, especially competion and query.